Refactoring CouchManager to use a configuration object to store URL, authentication info, etc...then, factor HttpClient out of CouchManager into a separate component to be reused in CouchChangesListener

+++ src/test/java/org/commonjava/web/maven/proxy/rest/access/GroupAccessResourceTest.java
@@ -42,7 +42,10 @@ public class GroupAccessResourceTest
     public static void clearRepoRootDir()
         throws IOException
     {
-        forceDelete( repoRoot );
+        if ( repoRoot != null && repoRoot.exists() )
+        {
+            forceDelete( repoRoot );
+        }
     }
 
     @Before
